Rename textinput -> text-input
